Move change handling code to WidgetFaram
- Remove un-necessary connect from redux
